Compilation Commit
This commit brings the application back into a compilable state.

- Added behavior tree includes in MahjongAIController.cpp (even though
they shouldn't be necessary).
- Switched uint32 to uint8 in MathBlueprintLibrary since uint32 is not
supported in Blueprints.
- Removed UPROPERTY from UMathBlueprintLibrary::SetPrecision_Double()
since double is not supported in Blueprints.
- Added missing MahjongAIController.h include in MahjongGameMode.cpp.
- Added GetGameMode() method to all GameModes.
- Added AMahjongCharacter.
- Added missing brace in MahjongPlayerController.cpp.
- Added stub Keybind fetch in MahjongSaveGame.
- Changed FindOrAdd to FindRef in MahjongSaveGame.cpp to keep
const-ness.
- Added Particle and Sound definitions to MahjongGame.h.
Changed incorrect access modifiers in GameMode classes.
- Added getters to MahjongPlayerState.h.

+++ Source/MahjongGame/Private/BlueprintFunctionLibraries/MathBlueprintLibrary.cpp
@@ -4,13 +4,13 @@
 #include "MathBlueprintLibrary.h"
 
 
-float UMathBlueprintLibrary::SetPrecision_Float(float Value, uint32 DecimalPlaces)
+float UMathBlueprintLibrary::SetPrecision_Float(float Value, uint8 DecimalPlaces)
 {
 	float PowerValue = FMath::Pow(10, DecimalPlaces);
 	return (float)(FMath::Floor(Value * PowerValue)) / PowerValue;
 }
 
-double UMathBlueprintLibrary::SetPrecision_Double(double Value, uint32 DecimalPlaces)
+double UMathBlueprintLibrary::SetPrecision_Double(double Value, uint8 DecimalPlaces)
 {
 	double PowerValue = FMath::Pow(10, DecimalPlaces);
 	return (double)(FMath::Floor(Value * PowerValue)) / PowerValue;
